Move all linting logic into lint.js
