+++ lib/lint.js
@@ -1,3 +1,9 @@
+const { relative } = require('path')
+const minimatch = require('minimatch')
+
+const findOptions = require('./findOptions')
+const getLinter = require('./getLinter')
+
 function getRange (line = 1, column = 1, source = '') {
   const line0 = line - 1
   const column0 = column - 1
@@ -7,16 +13,51 @@ function getRange (line = 1, column = 1, source = '') {
   return [start, end]
 }
 
-module.exports = function lint (linter, filePath, fileContent) {
-  return linter.lintText(filePath, fileContent)
-    .then(([{ messages } = {}]) => {
-      if (!Array.isArray(messages)) throw new Error('invalid lint report')
+function suppressError (err) {
+  return err.name === 'MissingLinterError' || err.name === 'MissingPackageError'
+}
 
-      return messages.map(({ fatal, message: text, line, column, source }) => ({
-        type: fatal ? 'Error' : 'Warning',
-        text,
-        filePath,
-        range: getRange(line, column, source)
-      }))
+const GRAMMAR_SCOPES = ['source.js', 'source.js.jsx']
+
+function lint (textEditor) {
+  const { scopeName } = textEditor.getGrammar()
+  if (!GRAMMAR_SCOPES.includes(scopeName)) return Promise.resolve([])
+
+  // Get text at the time the linter was invoked.
+  const fileContent = textEditor.getText()
+  const filePath = textEditor.getPath()
+
+  return findOptions(filePath)
+    .then(({ linterName, projectRoot, ignoreGlobs }) => {
+      const relativePath = relative(projectRoot, filePath)
+      const fileIsIgnored = ignoreGlobs.some(pattern => minimatch(relativePath, pattern))
+      if (fileIsIgnored) return [] // No errors
+
+      const linter = getLinter(linterName, projectRoot)
+      return linter.lintText(filePath, fileContent)
+        .then(([{ messages } = {}]) => {
+          if (!Array.isArray(messages)) throw new Error('invalid lint report')
+
+          return messages.map(({ fatal, message: text, line, column, source }) => ({
+            type: fatal ? 'Error' : 'Warning',
+            text,
+            filePath,
+            range: getRange(line, column, source)
+          }))
+        })
+    })
+    .catch(err => {
+      if (!suppressError(err)) {
+        atom.notifications.addError(err.message || 'Something bad happened', {
+          error: err,
+          detail: err.stack,
+          dismissable: true
+        })
+      }
+
+      return []
     })
 }
+
+exports = module.exports = lint
+exports.GRAMMAR_SCOPES = GRAMMAR_SCOPES
